:hammer: Move joypedal into a feature
